Refactoring sub-parsers (#93)
