+++ kit/command_init.py
@@ -148,3 +148,12 @@ def init_hekit(args):
     # Instructions for user
     print("Please, source your shell init file as follows")
     print(f"source {rc_file}")
+
+
+def set_init_subparser(subparsers, hekit_root_dir):
+    """create the parser for the 'init' command"""
+    parser_init = subparsers.add_parser("init", description="initialize hekit")
+    parser_init.add_argument(
+        "--default-config", action="store_true", help="setup default config file"
+    )
+    parser_init.set_defaults(fn=init_hekit, hekit_root_dir=hekit_root_dir)
