+++ kit/command_install.py
@@ -3,6 +3,7 @@
 
 """This module fetches, builds, or installs the requested libraries"""
 
+from typing import Dict
 from component_builder import components_to_build_from, chain_run
 
 
@@ -41,3 +42,37 @@ def install_components(args):
             continue
 
         chain_run(the_stages(component))
+
+
+def get_recipe_arg_dict(recipe_arg: str) -> Dict[str, str]:
+    """Returns a dictionary filled with recipe_arg values"""
+
+    pairs = [pair.split("=") for pair in recipe_arg.replace(" ", "").split(",")]
+    try:
+        return dict(pairs)
+    except ValueError as e:
+        for pair in pairs:
+            if len(pair) != 2:
+                raise ValueError(f"Wrong format for {pair}. Expected key=value") from e
+        return None
+
+
+def set_install_subparser(subparsers):
+    """create the parser for the 'install' command"""
+    actions = ["install", "build", "fetch"]
+
+    for action in actions:
+        parser = subparsers.add_parser(action, description=f"{action} components")
+        parser.add_argument(
+            "recipe_file",
+            metavar="recipe-file",
+            type=str,
+            help=f"TOML file for {action}",
+        )
+        parser.add_argument(
+            "--recipe_arg",
+            default={},
+            type=get_recipe_arg_dict,
+            help="Collection of key=value pairs separated by commas. The content of the TOML file will be replaced with this data.",
+        )
+        parser.set_defaults(fn=install_components, upto_stage=action)
