+++ kit/command_remove.py
@@ -5,6 +5,7 @@
 
 from shutil import rmtree
 from os import listdir
+from tab_completion import components_completer, instances_completer
 
 
 def remove_components(args):
@@ -27,3 +28,17 @@ def remove_components(args):
             "Nothing to remove",
             f"Instance '{instance}' of component '{component}' not found.",
         )
+
+
+def set_remove_subparser(subparsers):
+    """create the parser for the 'remove' command"""
+    parser_remove = subparsers.add_parser(
+        "remove", description="removes/uninstalls components"
+    )
+    parser_remove.add_argument(
+        "component", type=str, help="component to be removed"
+    ).completer = components_completer
+    parser_remove.add_argument(
+        "instance", type=str, help="instance to be removed"
+    ).completer = instances_completer
+    parser_remove.set_defaults(fn=remove_components)
