+++ tests/test_cmd_install.py
@@ -3,7 +3,7 @@
 
 import pytest
 from .context import command_install
-from command_install import install_components, _stages
+from command_install import install_components, _stages, get_recipe_arg_dict
 
 
 def test_install_components_all_unskipped(mocker, args, unskipped_components):
@@ -103,6 +103,57 @@ def test_stages_install(mocker, unskipped_components):
     assert next(act_result) == comp.fetch
     assert next(act_result) == comp.build
     assert next(act_result) == comp.install
+
+
+def test_get_recipe_arg_dict_correct_format():
+    """Arrange"""
+    act_arg = "key1=value1, key2=value2, key3=value3"
+    exc_dict = {"key1": "value1", "key2": "value2", "key3": "value3"}
+
+    """Act"""
+    act_dict = get_recipe_arg_dict(act_arg)
+
+    """Assert"""
+    assert exc_dict == act_dict
+
+
+def test_get_recipe_arg_dict_duplicated_key():
+    """Arrange"""
+    act_arg = "key1=value1, key1=value2, key3=value3"
+    exc_dict = {"key1": "value2", "key3": "value3"}
+
+    """Act"""
+    act_dict = get_recipe_arg_dict(act_arg)
+
+    """Assert"""
+    assert exc_dict == act_dict
+
+
+def test_get_recipe_arg_dict_wrong_format():
+    """Arrange"""
+    act_arg = "key1=value1, key1=value2, key3"
+
+    """Act"""
+    with pytest.raises(ValueError) as execinfo:
+        get_recipe_arg_dict(act_arg)
+
+    """Assert"""
+    assert "Wrong format for ['key3']. Expected key=value" == str(execinfo.value)
+
+
+def test_get_recipe_arg_dict_missing_comma():
+    """Arrange"""
+    act_arg = "key1=value1 key2=value2, key3=value3"
+
+    """Act"""
+    with pytest.raises(ValueError) as execinfo:
+        get_recipe_arg_dict(act_arg)
+
+    """Assert"""
+    assert (
+        "Wrong format for ['key1', 'value1key2', 'value2']. Expected key=value"
+        == str(execinfo.value)
+    )
 
 
 """Utilities used by the tests"""
